Separate authflow test files
