+++ e2e/tests/signup_test.yaml
@@ -8,10 +8,7 @@ steps:
         "name": "default"
       }
     assert:
-      - field: "type"
-        op: eq
-        value: "signup"
-      - field: "action.type"
+      - field: "result.action.type"
         op: eq
         value: "identify"
 
@@ -21,7 +18,7 @@ steps:
         "login_id": "signup@authgear.com"
       }
     assert:
-      - field: "action.type"
+      - field: "result.action.type"
         op: eq
         value: "verify"
 
@@ -30,7 +27,7 @@ steps:
         "code": "111111"
       }
     assert:
-      - field: "action.type"
+      - field: "result.action.type"
         op: eq
         value: "create_authenticator"
 
@@ -40,6 +37,6 @@ steps:
         "new_password":   "password"
       }
     assert:
-      - field: "action.type"
+      - field: "result.action.type"
         op: eq
         value: finished
